04/requirements.txt
Django==4.2
python-docx
nltk
scikit-learn==1.2.2
pandas